+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: berdogan <berdogan@student.42kocaeli.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/28 14:06:12 by berdogan          #+#    #+#             */
-/*   Updated: 2022/07/28 14:07:55 by berdogan         ###   ########.fr       */
+/*   Updated: 2022/07/28 16:16:32 by berdogan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,5 +20,5 @@
 # include<stdio.h>
 
 int	ft_printf(const char *src, ...);
-int	ft_itoa_base(unsigned long int nbr, int base, char format);
+int	ft_itoa_base(unsigned long long int nbr, int base, char format);
 #endif
